haproxy: per-template wait periods
The only template that needs to react to changes quickly is haproxy.tmpl
Blacklist and maintenance can be regenerated at a much lower rate than
the default 5s:30s.

As of consul-template 0.15.0, individual tempaltes can have their own
wait setting.  Dialing up the wait for maintenance and blacklist changes
_dramatically_ reduced load on the slaves, masters, and I/O utilization
on the masters.

https://app.datadoghq.com/dash/103411/berlioz-consul?live=false&page=0&is_auto=false&from_ts=1466748707766&to_ts=1469199783495&tile_size=m