registrator: turn off resync
after testing this in produciton for a week, we've seen zero drift
between service discovery and marathon

we turned this on agressively during the period where DD was DOSing the
docker daemons by collecting container sizes and it was necessary then.
Since then we've been gradually turning it down.  The resync
dramatically spikes consul/consul-template since it effectively removes
and re-adds all services.  Until that is addressed in registrator,
resync should be turned off.